rename server
